requirements.txt
flask
flask-sqlalchemy
flask-login
reportlab
gunicorn
